backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
# comtypes==1.4.9
contourpy==1.3.1
cssselect==1.3.0
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
Django==4.2.17
django-cors-headers==3.14.0
djangorestframework==3.15.2
djangorestframework-simplejwt==4.8.0
easyocr==1.7.2
executing==2.2.0
fastjsonschema==2.21.1
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
googletrans 
grpcio==1.68.1
gTTS==2.5.4
h11==0.16.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx 
huggingface-hub==0.27.0
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
ipykernel==6.29.5
# ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.7.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lxml==5.4.0
lxml_html_clean==0.4.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
ninja==1.11.1.4
nlp==0.4.0
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.2
numpy==2.0.2
openai==0.28.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.13.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.2
psutil==7.0.0
psycopg==3.2.6
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==18.1.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.21.1
python-json-logger==3.3.0
pyttsx3==2.98
pytz==2024.2
# pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.0
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
sacremoses==0.1.1
safetensors==0.4.5
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==75.6.0
sgmllib3k==1.0.0
shap==0.48.0
shapely==2.0.7
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
speech==0.5.2
SpeechRecognition==3.14.0
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
# tensorflow==2.18.0
termcolor==2.5.0
terminado==0.18.1
# tf_keras==2.18.0
threadpoolctl==3.6.0
tifffile==2025.3.13
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.0
xgboost==3.0.4
xxhash==3.5.0
yarl==1.18.3
